Changes to GetCurrentComposedLook method
Changes to GetCurrentComposedLook method, to work also in sites not in english. In English sites the current name of the theme is "Current" but in finnish sites ex. its "Nykyinen", so the method will fail. But if we use the id of the item, which is always 1, then we will get always the right theme.