update: Changes to GetCurrentComposedLook method
update: Changes to GetCurrentComposedLook method. If the site is not in english, there will not be item named "Current". for example in finnish sites it is "Nykyinen". But in composed looks list, the current item is always item number 1. So get the item id 1 from the list and you get always the currrent theme, no matter the language of the site.