Removing validation for PublishingWeb
On O365 an List that has been created using CSOM (in Hostweb) will end up throwing this error when trying to enable Ratings. 

But there seem to be no problem in creating Ratings in that list. So removing this Exception just makes it work.
MAYBE instead of removing that line, adding a "bool forcePublishingWeb=true" as additional Parameter might be even worth considering