Fix: if the principal is a sharepoint group, the name of the principal should also replace the site title as is done for sitegroups. If not an inconsistency can occur causing an exception when provisioning the template, the group/principal might not exist.
Fix: don't add a role assignment for a principal if the principal is a sharepoint group when sharepoint groups are not included. Provisioning will fail if the group/principal does not exist, and they will probably not exist (other than the default groups) if the are not in the provisioning template.

changed: only call ReplaceGroupTokens for sharepoint groups, otherwise there is no need for this method. This also prevents issues in rare cases where the loginName includes the site title or one of the default group names.