Updated to allow for the pageName to include relative path info.
When using OfficePnp Powershell and adding publishing pages (Add‑SPOPublishingPage), it is much easier to allow the relative path to be supplied as part of the page name parameter. The name parameter "folder/page.aspx" should provision a page in /Pages/folder/page.aspx path. Currently since the / is an invalid character we instead get /Pages/folder-page.aspx. Also since the / character is a standard separator seems like an intuitive addition.