Issue 1236
When fetching the formula of a calculated field using the Formula property the language would that of the lcid it was created with. When fetching the formula from the SchemaXml property we get the English/US version which will allow us to provision the formule when using ApplyProvisioningTemplate. See issue 1236 for more information.