Fix issue #1239 duplicate tokens
Using the web object in the token we can check if the token we are looking for is within the specified web. This code is only applied for provisioning a webpart, and I'm not sure if this would be usefull for other code that's dependant on the tokens. So for now I've provided seperate methods for only the webpart token parser: ParseXmlStringWebpart & ParseStringWebPart.