Fix issue #1239 & #1672: duplicate token webpart issue - only for webparts of type ContentBySearchWebPart
Replace the ~sitecollection prefix with the correct {sitecollection} token and make sure we keep the url after the ~sitecollection prefix. Using the web object in the token we can check if the token we are looking for is within the specified web. This code is only applied for provisioning a webpart, and I'm not sure if this would be usefull for other code that's dependant on the tokens. So for now I've provided seperate methods for only the webpart token parser: ParseXmlStringWebpart & ParseStringWebPart.