Update URL field value mapping for description.
As a requirement from one of our clients, it was needed to include a comma-separated text as the Description of the URL in our provisioning templates, but this was not possible due to the current code when uploading to SharePoint. This little change is just to be able to provision a URL field in the specified format (URL,Description) allowing the Description to contain commas, apart from any other character. With this change, urlArray will be always composed of two elements as long as there is a comma due to the required format.