Fixed an "InvalidOperationException"
... when extracting multilingual resources in a multi-threaded
environment. For example using a scheduler hosted in an ASP.NET
application, which executes multiple extraction jobs at the same time.

Due to the previously static "ResourceTokens" list an
"InvalidOperationException" with the message "Collection was modified;
enumeration operation may not execute" could occur. Now the already
available instance of "ProvisioningTemplateCreationInformation" hosts
the list and is used to add tokens to the list and store them within the
template.